Update LiveEnhancementSuite.ahk (Live 12 menu compatibility)
100ms was not enough when searching for Reverb.
The search results in Live 12 take a longer time when listing results for 'Reverb' so does not load the list in time to click on the first result. 

I have extended the time from 100ms to 200ms for each sleep before each Enter command.

I could whack it up to 500ms for greater compatibility on slower computers or bigger plugin libraries, but am now looking for a happy medium somewhere.

Perhaps a line in the settings.ini could be added in and linked to these values. But for now I've set it to 200ms each because 100 wasn't working.